get fork migration totals test to jest
